Move formatDuration to utils
